liquer-demo/requirements.txt
scikit-learn
Pygments
cryptography
pyarrow
fs
h5py
plotly
tornado
mkdocs
mkdocstrings
python-pptx
Pillow
datafusion
polars
fsspec
numpy>=1.15.4
pandas>=0.23.4
pytz>=2018.7
xlsxwriter>=1.1.2
pyparsing>=2.4.0
requests>=2.21.0
matplotlib>=3.0.2
sweetviz>=2.0.4
pyyaml
liquer-framework
liquer_pcv
liquer_gui